refactor: extract transition operators and payload builders from bending_tilt_leaflet.py
