+++ modules/energy/bending_tilt_leaflet.py
@@ -7,7 +7,6 @@ from typing import Dict
 import numpy as np
 
 from geometry.bending_derivatives import grad_triangle_area
-from geometry.curvature import compute_curvature_data
 from geometry.entities import Mesh
 from geometry.tilt_operators import (
     _resolve_transport_model,
@@ -21,12 +20,7 @@ from modules.energy.bending import (  # noqa: PLC0415
     _energy_model,
     _grad_cotan,
     _gradient_mode,
-    _spontaneous_curvature,
     _vertex_normals,
-)
-from modules.energy.leaflet_presence import (
-    leaflet_absent_vertex_mask,
-    leaflet_present_triangle_mask,
 )
 from modules.energy.scatter import scatter_triangle_scalar_to_vertices
 
@@ -39,14 +33,14 @@ from .bt_params import (
     _assume_J0_center_xy,
     _assume_J0_presets,
     _assume_J0_radius_max,
-    _base_term_boundary_group,
-    _base_term_region_mode,
-    _base_term_region_radius,
     _per_vertex_params_leaflet,
-    _resolve_bending_modulus,
     _use_inner_recovered_divergence,
     _use_stage_a_inner_shape_cross_suppression,
     _use_stage_a_outer_grad_linear_transition_operator,
+)
+from .bt_payload import (
+    _leaflet_static_tilt_payload,
+    _leaflet_triangle_payload,
 )
 from .bt_selection import (
     _apply_inner_divergence_update_mode,
@@ -56,800 +50,10 @@ from .bt_selection import (
     _interior_mask_leaflet,
     _shared_rim_support_transition_triangle_mask,
 )
-from .bt_utils import _accumulate_leaflet_tilt_gradient, _mean_reconstructed_field
-
-_OUTER_TRANSITION_OPERATOR_SHELLS = (
-    0.837822,
-    0.842474,
-    0.853008,
-    0.866643,
-    0.965910,
-    0.974541,
-    0.999984,
+from .bt_transition import (
+    _apply_transition_aware_beltrami_laplacian,
 )
-
-
-def _outer_transition_operator_payload(
-    mesh: Mesh,
-    *,
-    tri_rows_full: np.ndarray,
-    tri_rows: np.ndarray,
-) -> dict[str, np.ndarray]:
-    """Build cached transition-patch payload for the outer grad_linear operator."""
-    cache_attr = "_bending_tilt_out_transition_operator_payload"
-    cache_key = (
-        int(mesh._facet_loops_version),
-        int(mesh._vertex_ids_version),
-        int(tri_rows_full.shape[0]),
-        int(tri_rows.shape[0]),
-    )
-    cached = getattr(mesh, cache_attr, None)
-    if cached is not None and cached.get("key") == cache_key:
-        return cached["value"]
-
-    n_vertices = len(mesh.vertex_ids)
-    full_counts = np.zeros(n_vertices, dtype=np.int32)
-    keep_counts = np.zeros(n_vertices, dtype=np.int32)
-    for col in range(3):
-        np.add.at(full_counts, tri_rows_full[:, col], 1)
-        np.add.at(keep_counts, tri_rows[:, col], 1)
-
-    participation = np.zeros(n_vertices, dtype=float)
-    mask_full = full_counts > 0
-    participation[mask_full] = keep_counts[mask_full] / full_counts[mask_full]
-
-    domain = np.zeros(n_vertices, dtype=np.int8)
-    domain[(participation > 1.0e-12) & (participation < 1.0 - 1.0e-12)] = 1
-    domain[participation >= 1.0 - 1.0e-12] = 2
-
-    n_tri = tri_rows.shape[0]
-    edge_endpoint_a = np.empty((n_tri, 3), dtype=np.int32)
-    edge_endpoint_b = np.empty((n_tri, 3), dtype=np.int32)
-    edge_endpoint_a[:, 0] = tri_rows[:, 1]
-    edge_endpoint_b[:, 0] = tri_rows[:, 2]
-    edge_endpoint_a[:, 1] = tri_rows[:, 2]
-    edge_endpoint_b[:, 1] = tri_rows[:, 0]
-    edge_endpoint_a[:, 2] = tri_rows[:, 0]
-    edge_endpoint_b[:, 2] = tri_rows[:, 1]
-
-    edge_third = np.empty((n_tri, 3), dtype=np.int32)
-    edge_third[:, 0] = tri_rows[:, 0]
-    edge_third[:, 1] = tri_rows[:, 1]
-    edge_third[:, 2] = tri_rows[:, 2]
-
-    edge_to_thirds: dict[tuple[int, int], list[int]] = {}
-    for tri_idx in range(n_tri):
-        v0, v1, v2 = tri_rows[tri_idx]
-        edge_to_thirds.setdefault(tuple(sorted((int(v1), int(v2)))), []).append(int(v0))
-        edge_to_thirds.setdefault(tuple(sorted((int(v2), int(v0)))), []).append(int(v1))
-        edge_to_thirds.setdefault(tuple(sorted((int(v0), int(v1)))), []).append(int(v2))
-
-    recon_a_idx = np.full((n_tri, 3, 2), -1, dtype=np.int32)
-    recon_b_idx = np.full((n_tri, 3, 2), -1, dtype=np.int32)
-    recon_a_count = np.zeros((n_tri, 3), dtype=np.int8)
-    recon_b_count = np.zeros((n_tri, 3), dtype=np.int8)
-
-    for tri_idx in range(n_tri):
-        for edge_idx in range(3):
-            a = int(edge_endpoint_a[tri_idx, edge_idx])
-            b = int(edge_endpoint_b[tri_idx, edge_idx])
-            thirds = edge_to_thirds.get(tuple(sorted((a, b))), [])
-            a_candidates = [idx for idx in thirds if domain[idx] == domain[a]]
-            b_candidates = [idx for idx in thirds if domain[idx] == domain[b]]
-            if a_candidates:
-                recon_a_count[tri_idx, edge_idx] = min(len(a_candidates), 2)
-                recon_a_idx[tri_idx, edge_idx, : recon_a_count[tri_idx, edge_idx]] = (
-                    np.asarray(a_candidates[:2], dtype=np.int32)
-                )
-            if b_candidates:
-                recon_b_count[tri_idx, edge_idx] = min(len(b_candidates), 2)
-                recon_b_idx[tri_idx, edge_idx, : recon_b_count[tri_idx, edge_idx]] = (
-                    np.asarray(b_candidates[:2], dtype=np.int32)
-                )
-
-    edge_domain_a = domain[edge_endpoint_a]
-    edge_domain_b = domain[edge_endpoint_b]
-    transition_mask = ((edge_domain_a == 1) & (edge_domain_b == 2)) | (
-        (edge_domain_a == 2) & (edge_domain_b == 1)
-    )
-    same_mask = ~transition_mask
-
-    transition_full_vertex_mask = np.zeros(n_vertices, dtype=bool)
-    transition_full_a = transition_mask & (edge_domain_a == 2)
-    transition_full_b = transition_mask & (edge_domain_b == 2)
-    if np.any(transition_full_a):
-        transition_full_vertex_mask[np.unique(edge_endpoint_a[transition_full_a])] = (
-            True
-        )
-    if np.any(transition_full_b):
-        transition_full_vertex_mask[np.unique(edge_endpoint_b[transition_full_b])] = (
-            True
-        )
-    partial_vertex_mask = domain == 1
-    halo_full_vertex_mask = transition_full_vertex_mask & (domain == 2)
-    patch_vertex_mask = partial_vertex_mask | halo_full_vertex_mask
-
-    edge_unique_idx = np.empty((n_tri, 3), dtype=np.int32)
-    unique_edge_u: list[int] = []
-    unique_edge_v: list[int] = []
-    edge_to_unique: dict[tuple[int, int], int] = {}
-    for tri_idx in range(n_tri):
-        for edge_idx in range(3):
-            a = int(edge_endpoint_a[tri_idx, edge_idx])
-            b = int(edge_endpoint_b[tri_idx, edge_idx])
-            key = (a, b) if a < b else (b, a)
-            unique_idx = edge_to_unique.get(key)
-            if unique_idx is None:
-                unique_idx = len(unique_edge_u)
-                edge_to_unique[key] = unique_idx
-                unique_edge_u.append(key[0])
-                unique_edge_v.append(key[1])
-            edge_unique_idx[tri_idx, edge_idx] = unique_idx
-
-    unique_edge_u_arr = np.asarray(unique_edge_u, dtype=np.int32)
-    unique_edge_v_arr = np.asarray(unique_edge_v, dtype=np.int32)
-    patch_internal_edge_mask = (
-        patch_vertex_mask[unique_edge_u_arr] & patch_vertex_mask[unique_edge_v_arr]
-    )
-    patch_boundary_edge_mask = (
-        patch_vertex_mask[unique_edge_u_arr] ^ patch_vertex_mask[unique_edge_v_arr]
-    )
-    exterior_edge_mask = ~(patch_internal_edge_mask | patch_boundary_edge_mask)
-
-    adjacency: dict[int, list[int]] = {
-        int(row): [] for row in np.flatnonzero(patch_vertex_mask)
-    }
-    for edge_idx, is_internal in enumerate(patch_internal_edge_mask):
-        if not is_internal:
-            continue
-        u = int(unique_edge_u_arr[edge_idx])
-        v = int(unique_edge_v_arr[edge_idx])
-        adjacency[u].append(v)
-        adjacency[v].append(u)
-
-    patch_component_rows: list[np.ndarray] = []
-    patch_boundary_rows: list[np.ndarray] = []
-    seen = np.zeros(n_vertices, dtype=bool)
-    for start in np.flatnonzero(patch_vertex_mask):
-        if seen[start]:
-            continue
-        stack = [int(start)]
-        component: list[int] = []
-        seen[start] = True
-        while stack:
-            row = stack.pop()
-            component.append(row)
-            for neighbor in adjacency.get(row, ()):
-                if not seen[neighbor]:
-                    seen[neighbor] = True
-                    stack.append(neighbor)
-        component_rows = np.asarray(sorted(component), dtype=np.int32)
-        patch_component_rows.append(component_rows)
-
-        component_mask = np.zeros(n_vertices, dtype=bool)
-        component_mask[component_rows] = True
-        boundary_edge_mask = (
-            component_mask[unique_edge_u_arr] ^ component_mask[unique_edge_v_arr]
-        )
-        boundary_rows = np.unique(
-            np.concatenate(
-                (
-                    unique_edge_u_arr[
-                        boundary_edge_mask & ~component_mask[unique_edge_u_arr]
-                    ],
-                    unique_edge_v_arr[
-                        boundary_edge_mask & ~component_mask[unique_edge_v_arr]
-                    ],
-                )
-            )
-        )
-        patch_boundary_rows.append(boundary_rows.astype(np.int32, copy=False))
-
-    value = {
-        "domain": domain,
-        "participation": participation,
-        "edge_endpoint_a": edge_endpoint_a,
-        "edge_endpoint_b": edge_endpoint_b,
-        "edge_domain_a": edge_domain_a,
-        "edge_domain_b": edge_domain_b,
-        "edge_unique_idx": edge_unique_idx,
-        "unique_edge_u": unique_edge_u_arr,
-        "unique_edge_v": unique_edge_v_arr,
-        "recon_a_idx": recon_a_idx,
-        "recon_b_idx": recon_b_idx,
-        "recon_a_count": recon_a_count,
-        "recon_b_count": recon_b_count,
-        "transition_mask": transition_mask,
-        "same_mask": same_mask,
-        "partial_vertex_mask": partial_vertex_mask,
-        "halo_full_vertex_mask": halo_full_vertex_mask,
-        "transition_full_vertex_mask": transition_full_vertex_mask,
-        "patch_vertex_mask": patch_vertex_mask,
-        "patch_internal_edge_mask": patch_internal_edge_mask,
-        "patch_boundary_edge_mask": patch_boundary_edge_mask,
-        "exterior_edge_mask": exterior_edge_mask,
-        "patch_component_rows": patch_component_rows,
-        "patch_boundary_rows": patch_boundary_rows,
-    }
-    setattr(mesh, cache_attr, {"key": cache_key, "value": value})
-    return value
-
-
-def _apply_edge_reconstructed_beltrami_laplacian(
-    mesh: Mesh,
-    *,
-    positions: np.ndarray,
-    weights: np.ndarray,
-    tri_rows: np.ndarray,
-    tri_rows_full: np.ndarray,
-    field: np.ndarray,
-    cache_tag: str,
-    payload: dict[str, np.ndarray] | None = None,
-) -> tuple[np.ndarray, dict[str, object]]:
-    """Apply the merged PR #496 transition-edge reconstruction operator."""
-    if payload is None:
-        payload = _outer_transition_operator_payload(
-            mesh, tri_rows_full=tri_rows_full, tri_rows=tri_rows
-        )
-    edge_endpoint_a = np.asarray(payload["edge_endpoint_a"], dtype=np.int32)
-    edge_endpoint_b = np.asarray(payload["edge_endpoint_b"], dtype=np.int32)
-    recon_a_idx = np.asarray(payload["recon_a_idx"], dtype=np.int32)
-    recon_b_idx = np.asarray(payload["recon_b_idx"], dtype=np.int32)
-    recon_a_count = np.asarray(payload["recon_a_count"], dtype=np.int8)
-    recon_b_count = np.asarray(payload["recon_b_count"], dtype=np.int8)
-    transition_mask = np.asarray(payload["transition_mask"], dtype=bool)
-    same_mask = np.asarray(payload["same_mask"], dtype=bool)
-    participation = np.asarray(payload["participation"], dtype=float)
-    domain = np.asarray(payload["domain"], dtype=np.int8)
-
-    state_a = field[edge_endpoint_a]
-    state_b = field[edge_endpoint_b]
-    recon_a = _mean_reconstructed_field(
-        field, edge_endpoint_a, recon_a_idx, recon_a_count
-    )
-    recon_b = _mean_reconstructed_field(
-        field, edge_endpoint_b, recon_b_idx, recon_b_count
-    )
-
-    use_a = np.where(transition_mask[..., None], recon_a, state_a)
-    use_b = np.where(transition_mask[..., None], recon_b, state_b)
-    coeff = 0.5 * weights[:, :, None]
-    flux = coeff * (use_a - use_b)
-    transition_flux = np.where(transition_mask[..., None], flux, 0.0)
-    same_flux = np.where(same_mask[..., None], flux, 0.0)
-
-    out = np.zeros_like(field)
-    out_transition = np.zeros_like(field)
-    out_same = np.zeros_like(field)
-    for edge_idx in range(3):
-        a = edge_endpoint_a[:, edge_idx]
-        b = edge_endpoint_b[:, edge_idx]
-        f = flux[:, edge_idx, :]
-        ft = transition_flux[:, edge_idx, :]
-        fs = same_flux[:, edge_idx, :]
-        np.add.at(out, a, f)
-        np.add.at(out, b, -f)
-        np.add.at(out_transition, a, ft)
-        np.add.at(out_transition, b, -ft)
-        np.add.at(out_same, a, fs)
-        np.add.at(out_same, b, -fs)
-
-    radii = np.linalg.norm(positions[:, :2], axis=1)
-    shell_summary: list[dict[str, float | int]] = []
-    for target in _OUTER_TRANSITION_OPERATOR_SHELLS:
-        rows = np.where(np.isclose(radii, target, atol=1.0e-5))[0]
-        if rows.size == 0:
-            continue
-        shell_summary.append(
-            {
-                "radius": float(np.mean(radii[rows])),
-                "vertex_count": int(rows.size),
-                "transition_z_mean": float(np.mean(out_transition[rows, 2])),
-                "same_z_mean": float(np.mean(out_same[rows, 2])),
-                "total_z_mean": float(np.mean(out[rows, 2])),
-            }
-        )
-
-    transition_examples = np.argwhere(transition_mask)
-    example = None
-    if transition_examples.size:
-        tri_idx, edge_idx = transition_examples[0]
-        a = int(edge_endpoint_a[tri_idx, edge_idx])
-        b = int(edge_endpoint_b[tri_idx, edge_idx])
-        example = {
-            "triangle_index": int(tri_idx),
-            "edge_index": int(edge_idx),
-            "endpoint_rows": [a, b],
-            "endpoint_domains": [
-                "partial" if int(domain[a]) == 1 else "full",
-                "partial" if int(domain[b]) == 1 else "full",
-            ],
-            "endpoint_participation": [
-                float(participation[a]),
-                float(participation[b]),
-            ],
-            "raw_endpoint_factor_K_vec": [
-                field[a].tolist(),
-                field[b].tolist(),
-            ],
-            "reconstructed_side_states": [
-                recon_a[tri_idx, edge_idx].tolist(),
-                recon_b[tri_idx, edge_idx].tolist(),
-            ],
-        }
-
-    stats = {
-        "enabled": True,
-        "mode": "outer_grad_linear_transition_operator_v1",
-        "cache_tag": str(cache_tag),
-        "lane": str(mesh.global_parameters.get("theory_parity_lane") or ""),
-        "transition_edge_count": int(np.sum(transition_mask)),
-        "same_domain_edge_count": int(np.sum(same_mask)),
-        "nontransition_uses_raw_state": True,
-        "reconstructed_state_variable": "factor_K_vec",
-        "reconstruction_rule": "same-domain edge-adjacent third-vertex averaging",
-        "transition_example": example,
-        "shell_summary": shell_summary,
-    }
-    return out, stats
-
-
-def _apply_transition_aware_beltrami_laplacian(
-    mesh: Mesh,
-    *,
-    positions: np.ndarray,
-    weights: np.ndarray,
-    tri_rows: np.ndarray,
-    tri_rows_full: np.ndarray,
-    field: np.ndarray,
-    cache_tag: str,
-) -> tuple[np.ndarray, dict[str, object]]:
-    """Apply a patch-local outer grad_linear operator on the transition region."""
-    payload = _outer_transition_operator_payload(
-        mesh, tri_rows_full=tri_rows_full, tri_rows=tri_rows
-    )
-    patch_vertex_mask = np.asarray(payload["patch_vertex_mask"], dtype=bool)
-    if not np.any(patch_vertex_mask):
-        return _apply_edge_reconstructed_beltrami_laplacian(
-            mesh,
-            positions=positions,
-            weights=weights,
-            tri_rows=tri_rows,
-            tri_rows_full=tri_rows_full,
-            field=field,
-            cache_tag=cache_tag,
-            payload=payload,
-        )
-
-    edge_endpoint_a = np.asarray(payload["edge_endpoint_a"], dtype=np.int32)
-    edge_endpoint_b = np.asarray(payload["edge_endpoint_b"], dtype=np.int32)
-    edge_unique_idx = np.asarray(payload["edge_unique_idx"], dtype=np.int32)
-    unique_edge_u = np.asarray(payload["unique_edge_u"], dtype=np.int32)
-    unique_edge_v = np.asarray(payload["unique_edge_v"], dtype=np.int32)
-    transition_mask = np.asarray(payload["transition_mask"], dtype=bool)
-    participation = np.asarray(payload["participation"], dtype=float)
-    partial_vertex_mask = np.asarray(payload["partial_vertex_mask"], dtype=bool)
-    halo_full_vertex_mask = np.asarray(payload["halo_full_vertex_mask"], dtype=bool)
-    transition_full_vertex_mask = np.asarray(
-        payload["transition_full_vertex_mask"], dtype=bool
-    )
-    patch_internal_edge_mask = np.asarray(
-        payload["patch_internal_edge_mask"], dtype=bool
-    )
-    patch_boundary_edge_mask = np.asarray(
-        payload["patch_boundary_edge_mask"], dtype=bool
-    )
-    exterior_edge_mask = np.asarray(payload["exterior_edge_mask"], dtype=bool)
-    patch_component_rows = payload["patch_component_rows"]
-    patch_boundary_rows = payload["patch_boundary_rows"]
-
-    coeff_scalar = 0.5 * np.asarray(weights, dtype=float)
-    coeff_unique = np.bincount(
-        edge_unique_idx.ravel(),
-        weights=coeff_scalar.ravel(),
-        minlength=int(len(unique_edge_u)),
-    )
-
-    effective_field = field.copy()
-    fallback_component_count = 0
-    fallback_vertex_mask = np.zeros(len(mesh.vertex_ids), dtype=bool)
-    fallback = _apply_edge_reconstructed_beltrami_laplacian(
-        mesh,
-        positions=positions,
-        weights=weights,
-        tri_rows=tri_rows,
-        tri_rows_full=tri_rows_full,
-        field=field,
-        cache_tag=cache_tag,
-        payload=payload,
-    )
-    fallback_out = fallback[0]
-
-    for component_index, rows in enumerate(patch_component_rows):
-        rows = np.asarray(rows, dtype=np.int32)
-        if rows.size == 0:
-            continue
-        boundary_rows = np.asarray(patch_boundary_rows[component_index], dtype=np.int32)
-        if boundary_rows.size == 0:
-            fallback_component_count += 1
-            fallback_vertex_mask[rows] = True
-            continue
-
-        local_index = {int(row): idx for idx, row in enumerate(rows)}
-        size = int(rows.size)
-        lhs = np.zeros((size, size), dtype=float)
-        rhs = np.zeros((size, 3), dtype=float)
-        component_mask = np.zeros(len(mesh.vertex_ids), dtype=bool)
-        component_mask[rows] = True
-
-        touching_edges = np.where(
-            component_mask[unique_edge_u] | component_mask[unique_edge_v]
-        )[0]
-        for edge_idx in touching_edges:
-            weight = float(coeff_unique[edge_idx])
-            if weight == 0.0:
-                continue
-            u = int(unique_edge_u[edge_idx])
-            v = int(unique_edge_v[edge_idx])
-            iu = local_index.get(u, -1)
-            iv = local_index.get(v, -1)
-            if iu >= 0 and iv >= 0:
-                lhs[iu, iu] += weight
-                lhs[iv, iv] += weight
-                lhs[iu, iv] -= weight
-                lhs[iv, iu] -= weight
-            elif iu >= 0:
-                lhs[iu, iu] += weight
-                rhs[iu] += weight * field[v]
-            elif iv >= 0:
-                lhs[iv, iv] += weight
-                rhs[iv] += weight * field[u]
-
-        try:
-            solved = np.linalg.solve(lhs, rhs)
-        except np.linalg.LinAlgError:
-            fallback_component_count += 1
-            fallback_vertex_mask[rows] = True
-            if boundary_rows.size:
-                fallback_vertex_mask[boundary_rows] = True
-            continue
-        effective_field[rows] = solved
-
-    out = _apply_beltrami_laplacian(weights, tri_rows, effective_field)
-    if np.any(fallback_vertex_mask):
-        out[fallback_vertex_mask] = fallback_out[fallback_vertex_mask]
-
-    eff_a = effective_field[edge_endpoint_a]
-    eff_b = effective_field[edge_endpoint_b]
-    coeff = coeff_scalar[:, :, None]
-    flux = coeff * (eff_a - eff_b)
-    occ_internal_mask = patch_internal_edge_mask[edge_unique_idx]
-    occ_boundary_mask = patch_boundary_edge_mask[edge_unique_idx]
-    occ_exterior_mask = exterior_edge_mask[edge_unique_idx]
-
-    patch_internal_flux = np.where(occ_internal_mask[..., None], flux, 0.0)
-    patch_boundary_flux = np.where(occ_boundary_mask[..., None], flux, 0.0)
-    exterior_flux = np.where(occ_exterior_mask[..., None], flux, 0.0)
-
-    out_patch_internal = np.zeros_like(field)
-    out_patch_boundary = np.zeros_like(field)
-    out_exterior = np.zeros_like(field)
-    for edge_idx in range(3):
-        a = edge_endpoint_a[:, edge_idx]
-        b = edge_endpoint_b[:, edge_idx]
-        fi = patch_internal_flux[:, edge_idx, :]
-        fb = patch_boundary_flux[:, edge_idx, :]
-        fe = exterior_flux[:, edge_idx, :]
-        np.add.at(out_patch_internal, a, fi)
-        np.add.at(out_patch_internal, b, -fi)
-        np.add.at(out_patch_boundary, a, fb)
-        np.add.at(out_patch_boundary, b, -fb)
-        np.add.at(out_exterior, a, fe)
-        np.add.at(out_exterior, b, -fe)
-
-    radii = np.linalg.norm(positions[:, :2], axis=1)
-    patch_shell_summary: list[dict[str, float | int]] = []
-    for radius in np.unique(np.round(radii[patch_vertex_mask], 6)):
-        rows = np.where(patch_vertex_mask & np.isclose(radii, radius, atol=1.0e-5))[0]
-        if rows.size == 0:
-            continue
-        patch_shell_summary.append(
-            {
-                "radius": float(np.mean(radii[rows])),
-                "vertex_count": int(rows.size),
-                "participation_mean": float(np.mean(participation[rows])),
-            }
-        )
-
-    shell_summary: list[dict[str, float | int]] = []
-    for target in _OUTER_TRANSITION_OPERATOR_SHELLS:
-        rows = np.where(np.isclose(radii, target, atol=1.0e-5))[0]
-        if rows.size == 0:
-            continue
-        shell_summary.append(
-            {
-                "radius": float(np.mean(radii[rows])),
-                "vertex_count": int(rows.size),
-                "patch_internal_z_mean": float(np.mean(out_patch_internal[rows, 2])),
-                "patch_boundary_z_mean": float(np.mean(out_patch_boundary[rows, 2])),
-                "exterior_z_mean": float(np.mean(out_exterior[rows, 2])),
-                "total_z_mean": float(np.mean(out[rows, 2])),
-            }
-        )
-
-    patch_example = None
-    if patch_component_rows:
-        rows = np.asarray(patch_component_rows[0], dtype=np.int32)
-        boundary_rows = np.asarray(patch_boundary_rows[0], dtype=np.int32)
-        sample_rows = boundary_rows[: min(3, boundary_rows.size)]
-        patch_example = {
-            "component_index": 0,
-            "patch_vertex_count": int(rows.size),
-            "boundary_vertex_count": int(boundary_rows.size),
-            "component_radii": np.unique(np.round(radii[rows], 6)).tolist(),
-            "boundary_factor_K_vec_sample": effective_field[sample_rows].tolist(),
-        }
-
-    stats = {
-        "enabled": True,
-        "mode": "outer_grad_linear_transition_patch_operator_v1",
-        "cache_tag": str(cache_tag),
-        "lane": str(mesh.global_parameters.get("theory_parity_lane") or ""),
-        "transition_edge_count": int(np.sum(transition_mask)),
-        "same_domain_edge_count": int(np.sum(~transition_mask)),
-        "patch_component_count": int(len(patch_component_rows)),
-        "patch_vertex_count": int(np.sum(patch_vertex_mask)),
-        "partial_vertex_count": int(np.sum(partial_vertex_mask)),
-        "halo_full_vertex_count": int(np.sum(halo_full_vertex_mask)),
-        "halo_vertices_are_direct_transition_full_endpoints": bool(
-            np.array_equal(halo_full_vertex_mask, transition_full_vertex_mask)
-        ),
-        "fallback_component_count": int(fallback_component_count),
-        "exterior_uses_raw_state": True,
-        "reconstructed_state_variable": "factor_K_vec",
-        "reconstruction_rule": "patch-local Dirichlet harmonic solve",
-        "patch_example": patch_example,
-        "patch_shell_summary": patch_shell_summary,
-        "shell_summary": shell_summary,
-    }
-    return out, stats
-
-
-def _leaflet_triangle_payload(
-    mesh: Mesh,
-    global_params,
-    *,
-    positions: np.ndarray,
-    index_map: Dict[int, int],
-    cache_tag: str,
-    ctx=None,
-) -> dict[str, np.ndarray | None]:
-    """Return cached leaflet-masked triangle geometry for fixed positions."""
-    mesh.build_position_cache()
-    absent_mask = None
-    absent_key = None
-    if cache_tag in {"in", "out"}:
-        absent_mask = leaflet_absent_vertex_mask(mesh, global_params, leaflet=cache_tag)
-        absent_key = id(absent_mask)
-
-    use_cache = mesh._geometry_cache_active(positions)
-    cache_attr = f"_bending_tilt_leaflet_triangle_payload_cache_{cache_tag}"
-    cache_key = (
-        int(mesh._version),
-        int(mesh._facet_loops_version),
-        int(mesh._vertex_ids_version),
-        id(positions),
-        absent_key,
-        str(global_params.get("rim_slope_match_mode") or "")
-        if global_params is not None
-        else "",
-        str(global_params.get("rim_slope_match_outer_group") or "")
-        if global_params is not None
-        else "",
-        str(global_params.get("rim_slope_match_group") or "")
-        if global_params is not None
-        else "",
-        str(global_params.get("rim_slope_match_disk_group") or "")
-        if global_params is not None
-        else "",
-    )
-    if use_cache:
-        cached = getattr(mesh, cache_attr, None)
-        if cached is not None and cached.get("key") == cache_key:
-            payload = cached["value"]
-            if isinstance(payload, dict) and "tri_area" not in payload:
-                payload = dict(payload)
-                payload.setdefault("tri_area", None)
-                cached["value"] = payload
-            return payload
-
-    k_vecs, vertex_areas_vor, weights_full, tri_rows_full = compute_curvature_data(
-        mesh, positions, index_map
-    )
-    if tri_rows_full.size == 0:
-        payload = {
-            "k_vecs": k_vecs,
-            "vertex_areas_vor": vertex_areas_vor,
-            "weights_full": weights_full,
-            "tri_rows_full": tri_rows_full,
-            "weights": weights_full,
-            "tri_rows": tri_rows_full,
-            "tri_keep": np.zeros(0, dtype=bool),
-            "tri_area": None,
-            "g0": None,
-            "g1": None,
-            "g2": None,
-        }
-        if use_cache:
-            setattr(mesh, cache_attr, {"key": cache_key, "value": payload})
-        return payload
-
-    tri_keep = np.zeros(0, dtype=bool)
-    weights = weights_full
-    tri_rows = tri_rows_full
-    if absent_mask is not None:
-        tri_keep = leaflet_present_triangle_mask(
-            mesh, tri_rows_full, absent_vertex_mask=absent_mask
-        )
-        if tri_keep.size:
-            tri_rows = tri_rows_full[tri_keep]
-            weights = weights_full[tri_keep]
-
-    transition_mask = _shared_rim_support_transition_triangle_mask(
-        mesh,
-        global_params,
-        tri_rows_full,
-        keep_physical_outer_edge=str(cache_tag) == "out",
-    )
-    if transition_mask is not None:
-        keep_non_transition = ~transition_mask
-        if tri_keep.size:
-            tri_keep = tri_keep & keep_non_transition
-        else:
-            tri_keep = keep_non_transition
-        tri_rows = tri_rows_full[tri_keep]
-        weights = weights_full[tri_keep]
-
-    g0_use = None
-    g1_use = None
-    g2_use = None
-    area_use = None
-    if ctx is not None:
-        _area_cache, g0_cache, g1_cache, g2_cache, tri_rows_cache = (
-            ctx.geometry.p1_triangle_shape_gradients(mesh, positions)
-        )
-    else:
-        _area_cache, g0_cache, g1_cache, g2_cache, tri_rows_cache = (
-            mesh.p1_triangle_shape_gradient_cache(positions)
-        )
-    if tri_rows_cache.size and tri_rows_cache.shape[0] == tri_rows_full.shape[0]:
-        if tri_keep.size:
-            area_use = _area_cache[tri_keep]
-            g0_use = g0_cache[tri_keep]
-            g1_use = g1_cache[tri_keep]
-            g2_use = g2_cache[tri_keep]
-        else:
-            area_use = _area_cache
-            g0_use = g0_cache
-            g1_use = g1_cache
-            g2_use = g2_cache
-
-    payload = {
-        "k_vecs": k_vecs,
-        "vertex_areas_vor": vertex_areas_vor,
-        "weights_full": weights_full,
-        "tri_rows_full": tri_rows_full,
-        "weights": weights,
-        "tri_rows": tri_rows,
-        "tri_keep": tri_keep,
-        "tri_area": area_use,
-        "g0": g0_use,
-        "g1": g1_use,
-        "g2": g2_use,
-    }
-    if use_cache:
-        setattr(mesh, cache_attr, {"key": cache_key, "value": payload})
-    return payload
-
-
-def _leaflet_static_tilt_payload(
-    mesh: Mesh,
-    global_params,
-    *,
-    positions: np.ndarray,
-    index_map: Dict[int, int],
-    k_vecs: np.ndarray,
-    vertex_areas_vor: np.ndarray,
-    tri_rows: np.ndarray,
-    kappa_key: str,
-    cache_tag: str,
-) -> dict[str, np.ndarray]:
-    """Return cached fixed-geometry arrays used by tilt-only leaflet coupling."""
-    model = _energy_model(global_params)
-    if model != "helfrich":
-        model = "helfrich"
-    kappa_default = _resolve_bending_modulus(global_params, kappa_key)
-    c0_key = f"spontaneous_curvature_{cache_tag}"
-    c0_default = global_params.get(c0_key)
-    if c0_default is None:
-        c0_default = (
-            _spontaneous_curvature(global_params) if model == "helfrich" else 0.0
-        )
-    c0_default = float(c0_default or 0.0)
-
-    presets = _assume_J0_presets(global_params, cache_tag=cache_tag)
-    radius_max = _assume_J0_radius_max(global_params, cache_tag=cache_tag)
-    center_xy = _assume_J0_center_xy(global_params)
-    center_key = (float(center_xy[0]), float(center_xy[1]))
-    region_mode = _base_term_region_mode(global_params)
-    region_radius = _base_term_region_radius(global_params)
-    boundary_group = _base_term_boundary_group(global_params, cache_tag=cache_tag)
-
-    use_cache = mesh._geometry_cache_active(positions)
-    cache_attr = f"_bending_tilt_leaflet_static_cache_{cache_tag}"
-    cache_key = (
-        int(mesh._version),
-        int(mesh._facet_loops_version),
-        int(mesh._vertex_ids_version),
-        int(mesh._topology_version),
-        id(positions),
-        id(k_vecs),
-        id(vertex_areas_vor),
-        id(tri_rows),
-        str(model),
-        float(kappa_default),
-        float(c0_default),
-        None if boundary_group is None else str(boundary_group),
-        presets,
-        None if radius_max is None else float(radius_max),
-        center_key,
-        str(region_mode),
-        None if region_radius is None else float(region_radius),
-    )
-    if use_cache:
-        cached = getattr(mesh, cache_attr, None)
-        if cached is not None and cached.get("key") == cache_key:
-            return cached["value"]
-
-    kappa_arr, c0_arr = _per_vertex_params_leaflet(
-        mesh, global_params, model=model, kappa_key=kappa_key, cache_tag=cache_tag
-    )
-    safe_areas_vor = np.maximum(vertex_areas_vor, 1e-12)
-    k_mag = np.linalg.norm(k_vecs, axis=1)
-    h_vor = k_mag / (2.0 * safe_areas_vor)
-    is_interior = _interior_mask_leaflet(
-        mesh, global_params, cache_tag=cache_tag, index_map=index_map
-    )
-
-    base_term = (2.0 * h_vor) - c0_arr
-    base_term[~is_interior] = 0.0
-    if presets:
-        rows = _collect_preset_rows(
-            mesh,
-            presets=presets,
-            cache_tag=cache_tag,
-            index_map=index_map,
-            radius_max=radius_max,
-            center_xy=center_xy,
-        )
-        if rows.size:
-            base_term[rows] = 0.0
-    region_rows = _base_term_region_zero_rows(
-        mesh,
-        global_params,
-        cache_tag=cache_tag,
-        index_map=index_map,
-    )
-    if region_rows.size:
-        base_term[region_rows] = 0.0
-
-    value = {
-        "base_tri": base_term[tri_rows],
-        "kappa_tri": kappa_arr[tri_rows],
-    }
-    if use_cache:
-        setattr(mesh, cache_attr, {"key": cache_key, "value": value})
-    return value
+from .bt_utils import _accumulate_leaflet_tilt_gradient
 
 
 def _total_energy_leaflet(
